feat: API reservation calculation

+++ internal/models/reservation.go
@@ -11,10 +11,10 @@ type PersonalDataCalculation struct {
 }
 
 type SnacksCalculation struct {
-	ID       int    `json:"id"`
-	Category string `json:"category"`
-	Name     string `json:"name"`
-	Price    int    `json:"price"`
+	ID       int     `json:"id"`
+	Category string  `json:"category"`
+	Name     string  `json:"name"`
+	Price    float64 `json:"price"`
 }
 
 type RoomCalculation struct {
